Intl Grapheme: Adjust `grapheme_substr` PHP 8 compatibility
`grapheme_substr` is changed in [PHP 8.0 to return an empty string on out-of-bound offsets](https://php.watch/versions/8.0/substr-out-of-bounds#grapheme_substr).

This commit updates the tests with `@requires` annotations and polyfill updates to make sure the return results are in line with native function behavior. In PHP 8, an empty string (`""`) is returned if the length parameter is out of bounds of the given string. Prior to PHP 8.0, boolean `false` is returned.

Related: #288